Time_to_1st_exposed_in_GreenZone
Note: I had to copy the "SEPAIHRD_dynamics" files to a new directory on my computer to shorten the path. It was too long as it is and I could not open it in R for this reason (perhaps a problem linked to Microsoft computers, but I gotta work with what I have). If it is possible to make the names of the files and folders shorter, that could make my life easier.